Add files via upload
Using the same data set (credit_card_data.txt or credit_card_data-headers.txt) we use the ksvm or kknn function to find a good classifier:
(a) using cross-validation (k-nearest-neighbors model
(b) splitting the data into training, validation, and test data sets 

Then we use the iris data set iris.txt which contains 150 data points, each with four predictor variables and one categorical response. The predictors are the width and length of the sepal and petal of flowers and the response is the type of flower. The data is available from the R library datasets and can be accessed with iris once the library is loaded. It is also available at the UCI Machine Learning Repository (https://archive.ics.uci.edu/ml/datasets/Iris ). The response values are only given to see how well a specific method performed and should not be used to build the model. We use the R function kmeans to cluster the points as well as possible to report the best combination of 
predictors, suggested value of k, and how well the best clustering predicts flower type.