Add files via upload
Daily high temperature data for Atlanta for 1996 through 2015 for each year Using July through October were analyzed to estimate the "unofficial" end of summer. The raw data was reshaped from wide to long format and converted into a consistent date format. To explore, we applied exponential smoothing methods: 
1. Simple ES - smooths data using a level component only. 
2. Double ES (Holt's linear trend) - smooths with a level and trend component. 
3. Triple ES (Holt-Winters) - smooths with level, trend, and seasonal components.

Fitted values were extracted and compared visually and via mean squared error to determine which method captured the temperature patterns. Then the changepoint detection method utilizing the PELT algorithm on the smoothed daily temperatures to collect data on significant temperature drops for the second half of the year to avoid false fluctuations earlier in the summer. The data collected with the changepoint detection method was plotted to visualize trends in the end of summer across the 20 year period.