Add files via upload
US crime data and regression with the functions lm and glm are used to predict the observed crime rate in a city. The lm function fits a linear relationship between predictors 
and a response using ordinary least squares. The glm function allows us to specify the distribution fo the response and how predictors relate to response (link function). The 
following data is utilized for testing:
• M = 14.0
• So = 0
• Ed = 10.0
• Po1 = 12.0
• Po2 = 15.5
• LF = 0.640
• M.F = 94.0
• Pop = 150
• NW = 1.1
• U1 = 0.120
• U2 = 3.6
• Wealth = 3200
• Ineq = 20.1
• Prob = 0.04
• Time = 39.0