Add files via upload
US crime data is analyzed using:
a) A regression tree model
b) A random forest model.

The model quality is evaluated looking at the R2 and prediction results. For part a, the tree package is used to create a regression tree model. Then we prune the tree to highlight the most important splits and reduce variance and improve crossï¿¾validated performance. In the initial tree, we observe a split on Po1. Then on the left branch Pop and LF. On the right branch NW, Pop, and Po1. 

Then we use cross validation to determine the prune size to be used which is determined to be 7. However, when comparing the pruned trees, we observe that 5 nodes is less noisy and provides a simpler tree without losing accuracy and capturing variance. Part B uses the random forest and fits the crime data to a default 500 trees and 
bootstrapped sampling and random variable subsets for each split. 

Then in the second half, Logistic regression is used to find a good predictive model for whether credit applicants are good credit risks or not. The glm function in R is used to get a logistic regression (logit) model on data where the response is either zero or one, using family = binomial(link = ”logit”) in glm function call. In question 2, because the model gives a result between 0 and 1, it requires setting a threshold probability to separate between “good” and “bad” answers. In this data set, they estimate that incorrectly identifying a bad customer as good is 5 times worse than incorrectly classifying a good customer as bad. Therefore, we determine a good threshold 
probability based on the model.