Add files via upload
A regression model is built with US crime data using:
1. Stepwise regression
2. Lasso
3. Elastic net
For Parts 2 and 3, we first scale the data– otherwise, the regression coefficients will be on different scales, and the constraint won’t have the desired effect. For Parts 2 and 3, the glmnet function in R is used. For the elastic net model, λ is represented in glmnet as “alpha”; we obtain a range of results by varying alpha from 1 (lasso) to 0 (ridge regression) and, of course, other values of alpha in between. In a function call like glmnet(x, y, family=”mgaussian”, alpha=1) the predictors x needs to be in R’s matrix format, rather than data frame format. We convert a data frame to a matrix using as.matrix – for example, x <- as.matrix(data[,1:n-1]). Rather than specifying the value of T, glmnet returns models for a variety of values of T.